Pig-Stress-DP-Algorithm/requirements.txt
opencv-python 
numpy
pandas
matplotlib
seaborn
Django
djangorestframework
django-filter
markdown 
tensorflow==2.8.0
tensorflow-hub
tensorflow-gpu
tensorflow-io
pillow
scikit-learn
keras
wandb
jupyterlab
notebook
voila
albumentations
tqdm
urllib3
flask
flask-socketio
cmapy
RPI.GPIO
Adafruit-Blinka
adafruit-circuitpython-mlx90640
scipy>=1.6.0
pymongo
flask-cors